requirements.txt
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
streamlit>=1.28.0
scikit-learn>=1.3.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
requests>=2.31.0
